updated cashup page

+++ Backend/Controllers/CashupController.js
@@ -1,19 +1,32 @@
 import db from "../db.js";
 import Cashup from "../Models/Cashup.js";
+import { broadcastCustomerChanges } from "../SocketIO/socketIO.js";
 
 const CashupsCollection = db.collection("Cashups");
 
 // Create a new Cashup
 export const createCashup = async (req, res) => {
-  const { reasonName, profitOrOther, details, amount } = req.body;
+  const data = req.body;
 
   try {
-    const cashup = new Cashup(reasonName, profitOrOther, details, amount);
+    const cashup = new Cashup(
+      data.reasonName,
+      data.profitOrOther,
+      data.reasonDetails,
+      data.amount,
+      data.addedDateAndTime,
+      data.addedBy
+    );
 
     const docRef = await CashupsCollection.add({ ...cashup });
+
+    const newCashup = { id: docRef.id, ...cashup };
+
     res
       .status(201)
       .send({ message: "Cashup created successfully", id: docRef.id });
+
+    broadcastCustomerChanges("CashupAdded", newCashup);
   } catch (error) {
     res.status(500).send({ error: error.message });
   }
@@ -54,8 +67,19 @@ export const updateCashup = async (req, res) => {
   const updatedData = req.body;
 
   try {
+    const doc = await CashupsCollection.doc(id).get();
+    if (!doc.exists) {
+      return res.status(404).json({ message: "Cashup not found" });
+    }
+
     await CashupsCollection.doc(id).update(updatedData);
+
+    // Broadcast the updated customer to all POS systems
+    const updatedCashup = { id, ...updatedData };
+
     res.status(200).send({ message: "Cashup updated successfully" });
+
+    broadcastCustomerChanges("CashupUpdated", updatedCashup);
   } catch (error) {
     res.status(500).send({ error: error.message });
   }
@@ -66,8 +90,15 @@ export const deleteCashup = async (req, res) => {
   const { id } = req.params;
 
   try {
+    const doc = await CashupsCollection.doc(id).get();
+    if (!doc.exists) {
+      return res.status(404).json({ message: "Cashup not found" });
+    }
+
     await CashupsCollection.doc(id).delete();
     res.status(200).send({ message: "Cashup deleted successfully" });
+
+    broadcastCustomerChanges("CashupDeleted", { id });
   } catch (error) {
     res.status(500).send({ error: error.message });
   }
